+++ hax-lib/core-models/hax.sh
@@ -16,32 +16,17 @@ function extract_fstar() {
 function extract_lean() {
     go_to "./"
     LEAN_FILTERS=""
-    LEAN_FILTERS+=" -core_models::ops::function::Fn" # Issue #1710
-    LEAN_FILTERS+=" -core_models::result::**::ok" # Issue #1823
     LEAN_FILTERS+=" -core_models::result::**::unwrap" # Issue #1818
     LEAN_FILTERS+=" -core_models::result::**::expect" # Issue #1818
     LEAN_FILTERS+=" -core_models::option::**::expect" # Issue #1818
     LEAN_FILTERS+=" -core_models::option::**::unwrap" # Issue #1818
-    LEAN_FILTERS+=" -core_models::iter::traits::iterator::**" # Issue #1710
-    LEAN_FILTERS+=" -core_models::array::TryFromSliceError"
-    LEAN_FILTERS+=" -core_models::array::**::as_slice"
-    LEAN_FILTERS+=" -core_models::array::**::map"
-    LEAN_FILTERS+=" -core_models::array::**::from_fn"
-    LEAN_FILTERS+=" -core_models::array::iter::IntoIter"
-    LEAN_FILTERS+=" -core_models::Convert::impl_6"
-    LEAN_FILTERS+=" -core_models::mem::copy"
-    LEAN_FILTERS+=" -core_models::num::**::wrapping_add"
     LEAN_FILTERS+=" -core_models::num::**::saturating_add"
     LEAN_FILTERS+=" -core_models::num::**::overflowing_add"
-    LEAN_FILTERS+=" -core_models::num::**::wrapping_sub"
     LEAN_FILTERS+=" -core_models::num::**::saturating_sub"
     LEAN_FILTERS+=" -core_models::num::**::overflowing_sub"
-    LEAN_FILTERS+=" -core_models::num::**::wrapping_mul"
     LEAN_FILTERS+=" -core_models::num::**::saturating_mul"
     LEAN_FILTERS+=" -core_models::num::**::overflowing_mul"
-    LEAN_FILTERS+=" -core_models::num::**::pow"
     LEAN_FILTERS+=" -core_models::num::**::count_ones"
-    LEAN_FILTERS+=" -core_models::num::**::from_str_radix"
     LEAN_FILTERS+=" -core_models::num::**::rem_euclid"
     LEAN_FILTERS+=" -core_models::num::**::abs"
     LEAN_FILTERS+=" -core_models::num::**::checked_add"
@@ -50,33 +35,17 @@ function extract_lean() {
     LEAN_FILTERS+=" -core_models::num::**::MIN"
     LEAN_FILTERS+=" -core_models::num::**::MAX"
     LEAN_FILTERS+=" -core_models::num::**::BITS"
-    LEAN_FILTERS+=" -core_models::slice::iter::**"
-    LEAN_FILTERS+=" -core_models::slice::**::len"
-    LEAN_FILTERS+=" -core_models::slice::**::iter"
-    LEAN_FILTERS+=" -core_models::slice::**::is_empty"
-    LEAN_FILTERS+=" -core_models::slice::**::binary_search"
-    LEAN_FILTERS+=" -core_models::slice::**::copy_from_slice"
-    LEAN_FILTERS+=" -core_models::slice::**::clone_from_slice"
-    LEAN_FILTERS+=" -core_models::slice::**::chunks"
-    LEAN_FILTERS+=" -core_models::slice::**::chunks_exact"
-    LEAN_FILTERS+=" -core_models::slice::**::split_at"
-    LEAN_FILTERS+=" -core_models::slice::**::split_at_checked"
-    LEAN_FILTERS+=" -core_models::ops::deref::**"
-    LEAN_FILTERS+=" -core_models::iter::adapters::step_by::**"
-    LEAN_FILTERS+=" -core_models::iter::adapters::take::**"
-    LEAN_FILTERS+=" -core_models::iter::adapters::flat_map::**"
-    LEAN_FILTERS+=" -core_models::iter::adapters::flatten::**"
-    LEAN_FILTERS+=" -core_models::iter::adapters::zip::**"
-    LEAN_FILTERS+=" -core_models::iter::adapters::enumerate::**"
-    LEAN_FILTERS+=" -core_models::iter::adapters::map::**"
-    LEAN_FILTERS+=" -core_models::ops::range::**"
-    LEAN_FILTERS+=" -core_models::f32::**::abs"
+    LEAN_FILTERS+=" -core_models::num::**::from_be_bytes"
+    LEAN_FILTERS+=" -core_models::num::**::from_le_bytes"
+    LEAN_FILTERS+=" -core_models::num::**::to_be_bytes"
+    LEAN_FILTERS+=" -core_models::num::**::to_le_bytes"
+    LEAN_FILTERS+=" -core_models::num::**::rotate_left"
+    LEAN_FILTERS+=" -core_models::num::**::rotate_right"
     
     LEAN_FILTERS="$(echo "$LEAN_FILTERS" | xargs)"
     HAX_CORE_MODELS_EXTRACTION_MODE=on cargo hax into -i "$LEAN_FILTERS" lean
     sed -i 's/import Hax/import Hax.MissingCore/g' proofs/lean/extraction/Core_models.lean
     sed -i 's/def Ordering /def Ordering_ /g' proofs/lean/extraction/Core_models.lean # Issue #1646
-    sed -i 's/Core_models.Convert.From.from/Core_models.Convert.From._from/g' proofs/lean/extraction/Core_models.lean # Issue #1853
     
     cp proofs/lean/extraction/Core_models.lean ../proof-libs/lean/Hax/Core_models/Extracted.lean
 }